refactor: organize files
