+++ app/misc/entry-hattip.ts
@@ -6,8 +6,7 @@ import { SemanticAttributes } from "@opentelemetry/semantic-conventions";
 import * as build from "@remix-run/dev/server-build";
 import { createRequestHandler } from "@remix-run/server-runtime";
 import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
-import { requestContextStorageHandler } from "../server/request-context";
-import { responseHeadersContextHandler } from "../server/request-context-response-headers";
+import { requestContextHandler } from "../server/request-context";
 import { TRPC_ENDPOINT } from "../trpc/common";
 import { createTrpcAppContext } from "../trpc/context";
 import { trpcApp } from "../trpc/server";
@@ -23,8 +22,7 @@ export function createHattipEntry() {
     loggerMiddleware(),
     bootstrapHandler(),
     createTraceRequestHandler(),
-    requestContextStorageHandler(),
-    responseHeadersContextHandler(),
+    requestContextHandler(),
     rpcHandler(),
     createTrpchandler(),
     createRemixHandler()
