+++ app/server/README.md
@@ -1,0 +1,1 @@
+TODO: move server logic to this directory (e.g. entry-hattip.ts)
