+++ app/server/request-context/index.ts
@@ -1,0 +1,6 @@
+import { responseHeadersContextHandler } from "./response-headers";
+import { requestContextStorageHandler } from "./storage";
+
+export function requestContextHandler() {
+  return [requestContextStorageHandler(), responseHeadersContextHandler()];
+}
